+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- improve file detection during hydrate command for already present file in kustomization files
+
 ## [v2.0.1] - 2025-03-18
 
 ### Changed